requirements.txt

tensorflow>=2.12.0
numpy>=1.24.0
pillow>=9.0.0
customtkinter>=6.0.0
